Moved components to the component folder.
