+++ ts/Dashboards/Components/KPIComponent/KPISyncHandlers.ts
@@ -20,7 +20,7 @@
  * */
 
 import type DataCursor from '../../../Data/DataCursor';
-import type Sync from '../../Components/Sync/Sync';
+import type Sync from '../Sync/Sync';
 import type KPIComponent from './KPIComponent';
 
 import U from '../../../Core/Utilities.js';
